worker/requirements.txt
APScheduler==3.10.4
psycopg2-binary==2.9.9
docker==7.1.0
requests==2.32.3
python-dotenv==1.0.1
